WIP: working on fixing the histogram statistics Otsu calculus to find the split index. -- We're getting close, but there's still a wee problem exhibited by our magical-bug-trigger "Dance Troupe" test image:
Warning in pixOtsuAdaptiveThreshold: pix is ambivalent about whether black is foreground: weight < 1: -36 / 90, all-background don't-care count: 48

which is a new built-in warning for this kind of craziness: JPEG artifacts / noise impacting the Otsu threshold decision making.

Current status for the test image: almost all 'minority tiles' are flipped to majority behaviour without issue, except 3 of the, which became 'minority tiles' due to extra-dark JPEG noise, making the regular background *light* and thus black-fg minority in an otherwise white-fg test image. The key analysis bit to chew on is this: the threshold is below the median (obviously for almost-entirely-background tiles like the one at x=0,y=5). If we were to consider black-fg PIX *only*, things would become rather more simple: we could then simply *ignore* any valleys in the histogram to the right of the median, meaning we only would need to scan from the median on down for an Otsu-optimum valley.

Of course, things aren't as simple as that as there are several other tiles in the same boat, but not showing hallucinations like x=0,y=5 and x=2,j=5 do. So the idea is (!!TODO!!) to re-calc the Otsu threshold for minority tiles, starting at the median. While this would 'fix' x=0,y=5, it sure would nuke several other tiles in the test image which happen to be foreground-major, which is counter to the "background pixels are in the majority in every tile" assumption that underlays the whole thresholding kaboodle.

Ergo: hallucinations will remain, unless we raise our tile size to ensure the rule "background pixels are in the majority in every tile" is upheld for the entire PIX. (Which would fail permanently for very fat text a la the "KvK" test image (see the pixAutoPhotoInvert leptonica issue for the source and the problems they already have there, without this augmented histogram logic... TO BE CONTINUED...)

!!TODO!!: remove debug fprintf() calls.