Alternative solution for adaptive Otsu when some tiles ave a minority opinion about 'black is background'. This time, we inspect all tiles once the threshold calculus is done: any minority item will be turned into a majority opinion one by a special adjustment process: - first, the tile's histogram is *patched* to ensure the 'background' side is a simplified single hump (placed at the median index), so that the next step will not ever be able to make the same decision mistake as it did before: - next, the histogram is inspected as usual and an otsu threshold is obtained (if available): thanks to the first patch step above, this threshold will always at the majority side of things.
As we now have a suitable threshold map AND a global majority opinion about whether 'black is foreground', we now execute the threshold operation itself, but produce INVERTED output when 'black is BACKGROUND': the otsu result is now ALWAYS an image where black is foreground!

This also fixes the fringe condition where some tiles might be all-background with high threshold value (255).