WIP: further work on diagspec support and propagation: diagspec is attached to each PIX and the intent is to have the diagspec travel along from input to output PIX, while the image(s) are transformed/processed.
Also checked most of the rest of the lib to ensure those APIs that need a 'debugflag' get one; meanwhile a rule of thumb now is: if an API receives a source PIX, we don't need a debugflag or diagspec argument as well, as those are both embedded in the PIX!

The design/concept with the diagspec APIs is that those MUST tolerate NULL valued LDIAG_CTX arguments, so that the rest of the code can use the simplest possible usage of these APIs: minimal cluttering of other parts of the codebase is the intent. A preprocessor setting in utils4.c is set to 1 to allow extra *warnings* to be logged while we are still debugging/trailing this API set.