ABORTED: propagating diagspec to genPathname() (where it is needed if we want to have full control over tmp/out file path generation AND add file I/O statistics gathering, as we do) is turning out too cumbersome: it deteriorates the API's readability and consequently the code readability (see the /prog/ sources).
The consequence is that the only thing that will work is a GLOBAL single diagspec instance, if we do not want to clutter the function interfaces.
Alas, that reduces the 'propagate diagspec in PIX' work done previously to a moot point and useless activity.

Given this conclusion, that activity will be rolled back in the master development branch.

**END**