REVERTING the 'attach diagspec to PIX et al' work done before: see also SHA-1: 9bf63307621e2d3a6a0f1c21ed2cb2f6a88e1977 for the reason why.
Unfortunately this looses us the ability to 'debug per *specific PIX*', but it nicely declutters the leptonica API and codebase: we DO NOT re-introduce the old (mainline/vanilla leptonica) debug flags in some of those APIs, but rely on the global APIs throughout.

TODO: further work on the /tmp/ directory expansion/rewriting for safe behaviour that's capable of dealing with bulk runs/testing of the leptonica library in debug mode, when (a lot of) extra files are generated as part of the debugging process.

---

Quoting SHA-1: 9bf63307621e2d3a6a0f1c21ed2cb2f6a88e1977:

* ABORTED: propagating diagspec to genPathname() (where it is needed if we want to have full control over tmp/out file path generation AND add file I/O statistics gathering, as we do) is turning out too cumbersome: it deteriorates the API's readability and consequently the code readability (see the /prog/ sources).

The consequence is that the only thing that will work is a GLOBAL single diagspec instance, if we do not want to clutter the function interfaces.
Alas, that reduces the 'propagate diagspec in PIX' work done previously to a moot point and useless activity.

Given this conclusion, that activity will be rolled back in the master development branch.