gplot APIs: use the file path generated for gplot to write its output to as the sourcefile path for our `pixRead()` call which is meant to load those gplot charts into leptonica memory.
The goal of this (and other changes) is to arrive at a situation where leptonica does not ever need to use fixed paths, not even when regression testing. And the generated output & temporary file paths should strive to be unique enough to survive in a batch running environment for both test and production scenarios.